Button class rename + builder/static factory
