+++ src/main/java/com/n1netails/n1netails/slack/validation/impl/element/ButtonElementValidator.java
@@ -1,12 +1,12 @@
 package com.n1netails.n1netails.slack.validation.impl.element;
 
 import com.n1netails.n1netails.slack.exception.SlackValidationException;
-import com.n1netails.n1netails.slack.model.actions_element.Button;
+import com.n1netails.n1netails.slack.model.actions_element.ButtonElement;
 import com.n1netails.n1netails.slack.validation.SlackValidator;
 
-public class ButtonElementValidator implements SlackValidator<Button> {
+public class ButtonElementValidator implements SlackValidator<ButtonElement> {
     @Override
-    public void validate(Button target) throws SlackValidationException {
+    public void validate(ButtonElement target) throws SlackValidationException {
         if (target.getText() == null || target.getText().isBlank()) {
             throw new SlackValidationException("Button text is required");
         }
